+++ s2_fish/plot_solver_periodic.py
@@ -44,9 +44,9 @@ print "init =",t_init,", deposit =",t_deposit,", solve =",t_solve
 num_points = n
 ax = numpy.arrayrange(num_points)*size[0]/(num_points -1) - 0.5*size[0]
 az = numpy.arrayrange(num_points)*size[2]/(num_points -1) - 0.5*size[2]
-aphi = numpy.zeros([num_points],numpy.Float)
-arho = numpy.zeros([num_points],numpy.Float)
-aexact = numpy.zeros([num_points],numpy.Float)
+aphi = numpy.zeros([num_points],numpy.float64)
+arho = numpy.zeros([num_points],numpy.float64)
+aexact = numpy.zeros([num_points],numpy.float64)
 index = 0;
 for k in range(0,len(ax)):
     z = az[k]
